Move PerRegionFeature to own base-classes folder, and remove some unused imports.
