+++ src/ogd/games/PENGUINS/features/EggRecoverTime.py
@@ -5,8 +5,6 @@ from typing import Any, Dict, List, Optional
 from ogd.core.utils.Logger import Logger
 from datetime import datetime, timedelta
 from ogd.core.generators.Generator import GeneratorParameters
-from ogd.core.generators.extractors.Feature import Feature
-from ogd.games.PENGUINS.features.PerRegionFeature import PerRegionFeature
 from ogd.core.schemas.Event import Event
 from ogd.core.schemas.ExtractionMode import ExtractionMode
 from ogd.core.schemas.FeatureData import FeatureData
