+++ src/ogd/games/PENGUINS/features/WaddleCount.py
@@ -4,8 +4,6 @@ from typing import Any, Dict, List, Optional
 # import locals
 from ogd.core.utils.Logger import Logger
 from ogd.core.generators.Generator import GeneratorParameters
-from ogd.core.generators.extractors.Feature import Feature
-from ogd.games.PENGUINS.features.PerRegionFeature import PerRegionFeature
 from ogd.core.schemas.Event import Event
 from ogd.core.schemas.ExtractionMode import ExtractionMode
 from ogd.core.schemas.FeatureData import FeatureData
